refactor: move profile screen to own folder
