+++ src/modules/brand/screens/MyBrand/MyBrand.tsx
@@ -17,7 +17,7 @@ import { Avatar } from 'modules/profile/components/Avatar';
 import { Header } from 'modules/profile/components/Header';
 import { InfoPanel } from 'modules/profile/components/InfoPanel';
 import { SetBgImgModal } from 'modules/profile/components/SetBgImgModal';
-import { useProfileStyles } from 'modules/profile/screens/useProfileStyles';
+import { useProfileStyles } from 'modules/profile/screens/Profile/useProfileStyles';
 import { Section } from 'modules/uiKit/Section';
 import React, { useCallback, useEffect, useState } from 'react';
 import { useParams } from 'react-router';
@@ -37,12 +37,17 @@ export const MyBrand = () => {
 
   useEffect(() => {
     if (address && id) {
-      dispatch(queryBrandById({
-        id: parseInt(id),
-        accountaddress: address
-      }, {
-        asMutation: true,
-      })).then(res => {
+      dispatch(
+        queryBrandById(
+          {
+            id: parseInt(id),
+            accountaddress: address,
+          },
+          {
+            asMutation: true,
+          },
+        ),
+      ).then(res => {
         const brandInfo = res.data;
         if (brandInfo) {
           setBrandInfo(brandInfo);
@@ -111,12 +116,16 @@ export const MyBrand = () => {
                 href={
                   item.poolId && item.poolType
                     ? BuyNFTRoutesConfig.DetailsNFT.generatePath(
-                      item.poolId,
-                      item.poolType,
-                    )
+                        item.poolId,
+                        item.poolType,
+                      )
                     : ''
                 }
-                price={item.poolId && item.price ? new BigNumber(Web3.utils.fromWei(item.price)) : undefined}
+                price={
+                  item.poolId && item.price
+                    ? new BigNumber(Web3.utils.fromWei(item.price))
+                    : undefined
+                }
                 copies={item.supply}
                 MediaProps={{
                   category: item.category,
@@ -130,8 +139,9 @@ export const MyBrand = () => {
                   users: [
                     {
                       name: 'name',
-                      avatar: `${profileInfo?.imgUrl ?? 'https://via.placeholder.com/32'
-                        }`,
+                      avatar: `${
+                        profileInfo?.imgUrl ?? 'https://via.placeholder.com/32'
+                      }`,
                       verified: true,
                     },
                   ],
