+++ src/modules/i18n/en-US.json
@@ -286,6 +286,7 @@
     },
     "tabs": {
       "my-items": "My items",
+      "my-bids": "My bids",
       "my-brands": "My brands",
       "activity": "Activity",
       "liked": "Liked",
