+++ src/modules/profile/screens/Profile/components/TabItems/index.ts
@@ -1,0 +1,1 @@
+export * from './TabItems';
