+++ src/modules/profile/screens/Profile/index.ts
@@ -1,0 +1,1 @@
+export * from './Profile';
